OD20DEVAT3-33. Split responsibility of ServerManger to few classes. Minor cleanup.
